+++ src/components/home/Home.test.tsx
@@ -9,8 +9,8 @@ import { Home } from '@/pages/home';
 import { mockedUseNavigate } from '@/shared/test/setupTests';
 import { getTimeColor } from '@/shared/utils/getTimeColor';
 import userEvent from '@testing-library/user-event';
-import { useGetNotifications } from '../notification/queries';
 import LayoutWithNav from '../navigation/LayoutWithNav';
+import { useGetNotifications } from '@/features/notification/model';
 
 // vi.mock을 사용해 특정 모듈을 모킹할 수 있다.
 // 실제로 useGetBestAuctions 함수를 실행하는 대신, 원하는 반환값을 제공하는 모의 함수를 제공한다는 뜻
