+++ src/features/bid/api/index.ts
@@ -1,0 +1,1 @@
+export { postBid } from './postBid';
