+++ src/features/bid/api/postBid.ts
@@ -1,0 +1,7 @@
+import { API_END_POINT, httpClient } from '@/shared';
+import type { IBidPostData } from '../config';
+
+export const postBid = async (bidData: IBidPostData) => {
+  await httpClient.post(API_END_POINT.BID, bidData);
+  return;
+};
