+++ src/features/bid/index.ts
@@ -1,0 +1,1 @@
+export { BidForm } from './ui';
