+++ src/features/bid/model/index.ts
@@ -1,0 +1,1 @@
+export { usePostBid } from './usePostBid';
