+++ src/features/bid/model/usePostBid.ts
@@ -1,11 +1,10 @@
 import { UseMutateFunction, useMutation } from '@tanstack/react-query';
 
-import type { IBidPostData } from '@/features/bid/config/type';
-import { httpClient } from '@/shared/api/axios';
-import { API_END_POINT } from '@/shared/constants/apiEndPoint';
-import { ROUTES } from '@/shared/constants/routes';
+import { ROUTES } from '@/shared';
 import { useNavigate } from 'react-router-dom';
 import { toast } from 'sonner';
+import { postBid } from '../api';
+import { IBidPostData } from '../config';
 
 export const usePostBid = (
   auctionId: number
@@ -13,11 +12,6 @@ export const usePostBid = (
   mutate: UseMutateFunction<void, Error, IBidPostData, unknown>;
   isPending: boolean;
 } => {
-  const postBid = async (bidData: IBidPostData) => {
-    await httpClient.post(API_END_POINT.BID, bidData);
-    return;
-  };
-
   const navigate = useNavigate();
   const { mutate, isPending } = useMutation({
     mutationFn: postBid,
