+++ src/features/bid/test/EditBid.test.tsx
@@ -5,7 +5,7 @@ import { describe, expect, test, vi } from 'vitest';
 import { Bid } from '@/pages/bid/ui/Bid';
 import { mockedUseNavigate } from '@/shared/test/setupTests';
 import userEvent from '@testing-library/user-event';
-import { useGetAuctionDetails } from '../details/queries';
+import { useGetAuctionDetails } from '../../../components/details/queries';
 
 vi.mock('@/components/details/queries');
 vi.mocked(useGetAuctionDetails).mockReturnValue({
