+++ src/features/bid/ui/index.ts
@@ -1,0 +1,2 @@
+export { BidCaution } from './BidCaution';
+export { BidForm } from './BidForm';
