+++ src/features/notification/api/deleteNotification.ts
@@ -1,0 +1,6 @@
+import { API_END_POINT, httpClient } from '@/shared';
+
+export const deleteNotification = async (id: number) => {
+  await httpClient.delete(`${API_END_POINT.NOTIFICATIONS}/${id}`);
+  return;
+};
