+++ src/features/notification/test/Notification.test.tsx
@@ -1,16 +1,12 @@
 import { render, screen, waitFor } from '@testing-library/react';
 import { MemoryRouter, Route, Routes } from 'react-router-dom';
 import { describe, expect, test, vi } from 'vitest';
-import {
-  useDeleteNotification,
-  useGetNotifications,
-  useReadNotification,
-} from '../../../components/notification/queries';
 
 import { notificationData } from '@/mocks/data/notificationData';
 import { Notification } from '@/pages/notification';
 import { mockedUseNavigate } from '@/shared/test/setupTests';
 import userEvent from '@testing-library/user-event';
+import { useDeleteNotification, useGetNotifications, useReadNotification } from '..';
 import LayoutWithNav from '../../../components/navigation/LayoutWithNav';
 
 vi.mock('@/components/notification/queries', () => ({
