+++ src/features/notification/ui/index.ts
@@ -1,0 +1,2 @@
+export { RealTimeNotification } from './RealTimeNotification';
+export { RealTimeNotificationItem } from './RealTimeNotificationItem';
