+++ src/pages/notification/ui/Notification.tsx
@@ -1,11 +1,10 @@
-import NotificationItem from '@/components/notification/NotificationItem';
-import { useGetNotificationsWithSuspense } from '@/components/notification/queries';
-import type { INotification } from '@/features/notification/config/type';
+import { useGetNotificationsWithSuspense } from '@/features/notification';
+import type { INotification } from '@/features/notification/config';
+import { NotificationItem } from '@/features/notification/ui/NotificationItem';
 import { EmptyBoundary } from '@/shared';
 
 export const Notification = () => {
   const { notifications } = useGetNotificationsWithSuspense();
-
   return (
     <EmptyBoundary length={notifications.length} name='notification'>
       <div className='mx-[-32px] my-[-16px]'>
