+++ src/shared/index.ts
@@ -1,3 +1,4 @@
+export * from './api';
 export * from './constants/index';
 export * from './ui/index';
 export * from './utils/index';
